Enemy AI
Enemies head towards player.
There is a a problem with the behaviour where sometimes the enemies go
in the other direction. This is because of math when either the
player/enemy is around 345 - 360 degrees and the other is around 0-15
degrees. will need some if statements to alter it.
NOTE: watching movies does not help with writing code haha